Remove Ping method from interface and class
The `Ping` method has been removed from both the `IOfficeUtilsExternalLib` interface and the `OfficeUtilsExternalLib` class. This method was previously annotated with `[OSAction]` and described as an action to prevent AWS Lambda "cold starts" by running within a Timer. The removal indicates that this functionality is no longer needed or has been refactored elsewhere.